Added decimal point and digit to float/double literals

+++ Views/Panels/SplashScreenPanel.swift
@@ -2,11 +2,11 @@ import SwiftUI
 
 struct SplashScreenPanel: View {
     @Binding var isOpen: Bool
-    var duration: UInt64 = 15
+    var duration: UInt64 = UInt64(15)
     var isWide: Bool
     
     var body: some View {
-        BlueprintGrid(baseSpacing: 64, lineWidth: isWide ? 1.5 : 0.75)
+        BlueprintGrid(baseSpacing: 64.0, lineWidth: isWide ? 1.5 : 0.75)
             .task(priority: TaskPriority.medium, delayCover)
             .zIndex(1)
             .scaleEffect(isWide ? 2.76 : 4.8)
@@ -22,11 +22,11 @@ struct SplashScreenPanel: View {
                             Styling.appIcon.swuiImage
                                 .rs(fit: true)
                                 .mask(Styling.roundedRect)
-                                .shadow(color: Styling.black.opacity(0.75), radius: 10)
+                                .shadow(color: Styling.black.opacity(0.75), radius: 10.0)
                             Text("Open Brick Art Instructor").font(Styling.titleFont)
                             Text("(will continue in \(duration)s)").font(Styling.caption2Font)
                         }
-                        .frameRow(300, Alignment.center)
+                        .frameRow(300.0, Alignment.center)
                     })
                 })
                 .padding()
